(#181) Integrate Console::History with the Console itself

+++ src/something_commands.cpp
@@ -93,7 +93,7 @@ void command_reload(Game *game, String_View)
 void command_history(Game *game, String_View)
 {
     game->console.println("--------------------");
-    for (size_t i = 0; i < game->console.history.count; ++i) {
+    for (int i = 0; i < game->console.history.count; ++i) {
         const size_t j = (game->console.history.begin + i) % CONSOLE_HISTORY_CAPACITY;
         const String_View entry = {
             game->console.history.entry_sizes[j],
